Debug Mode (#122)
* Added test command line option and wrote scripts

* Added CLI11

* Fix Config

* Testing off

* Changed name of testing flag

* Added debug flag

* Added LOG_USER

* Added timeout for integration tests

* install timeout library

* Added integration test artifacts to upload

* fix CI

* Switch to v2

* Change timeout to 5 sec

* Absolute path

* defaulted debug to false

+++ include/List.hpp
@@ -209,7 +209,7 @@ public:
         std::stringstream stream;
         for (uint32_t i = 0; i < m_length; i++)
         {
-            stream << m_list[i];
+            stream << static_cast<uint32_t>(m_list[i]);
             if (i != m_length - 1)
             {
                 stream << ", ";
